+++ spec/spec/line.js
@@ -112,6 +112,18 @@ describe('Line', function() {
       expect(box.x + box.width).toBe(987)
       expect(box.y).toBe(0)
     })
+    it('defines the width and height proportionally with only the width value given', function() {
+      var box = line.bbox()
+      line.size(500)
+      expect(line.width()).toBe(500)
+      expect(line.width() / line.height()).toBe(box.width / box.height)
+    })
+    it('defines the width and height proportionally with only the height value given', function() {
+      var box = line.bbox()
+      line.size(null, 525)
+      expect(line.height()).toBe(525)
+      expect(line.width() / line.height()).toBe(box.width / box.height)
+    })
   })
   
   describe('scale()', function() {