+++ spec/spec/path.js
@@ -102,11 +102,23 @@ describe('Path', function() {
   })
   
   describe('size()', function() {
-    it('should define the width and height of the element', function() {
+    it('defines the width and height of the element', function() {
       path.size(987,654)
       var box = path.bbox()
       expect(approximately(box.width, 0.1)).toBe(987)
       expect(approximately(box.height, 0.1)).toBe(654)
+    })
+    it('defines the width and height proportionally with only the width value given', function() {
+      var box = path.bbox()
+      path.size(500)
+      expect(path.width()).toBe(500)
+      expect(path.width() / path.height()).toBe(box.width / box.height)
+    })
+    it('defines the width and height proportionally with only the height value given', function() {
+      var box = path.bbox()
+      path.size(null, 525)
+      expect(path.height()).toBe(525)
+      expect(path.width() / path.height()).toBe(box.width / box.height)
     })
   })
   
@@ -137,7 +149,7 @@ describe('Path', function() {
   describe('plot()', function() {
     it('falls back to a single point without an argument', function() {
       path = draw.path()
-      expect(path.node.getAttribute('d')).toBe('M0,0')
+      expect(path.node.getAttribute('d')).toBe('M 0 0')
     })
   })