+++ src/path.js
@@ -1,75 +1,49 @@
-SVG.Path = function(unbiased) {
+SVG.Path = function() {
   this.constructor.call(this, SVG.create('path'))
-  
-  this.unbiased = !!unbiased
 }
 
 // Inherit from SVG.Shape
 SVG.Path.prototype = new SVG.Shape
 
 SVG.extend(SVG.Path, {
-  // Move over x-axis
-  x: function(x) {
-    return x == null ? this.bbox().x : this.transform('x', x)
+  // Plot new poly points
+  plot: function(p) {
+    return this.attr('d', (this.array = new SVG.PathArray(p, [{ type:'M',x:0,y:0 }])))
   }
-  // Move over y-axis
+  // Move by left top corner
+, move: function(x, y) {
+    return this.attr('d', this.array.move(x, y))
+  }
+  // Move by left top corner over x-axis
+, x: function(x) {
+    return x == null ? this.bbox().x : this.move(x, this.bbox().y)
+  }
+  // Move by left top corner over y-axis
 , y: function(y) {
-    return y == null ? this.bbox().y : this.transform('y', y)
+    return y == null ? this.bbox().y : this.move(this.bbox().x, y)
+  }
+  // Set element size to given width and height
+, size: function(width, height) {
+    var p = this._proportionalSize(width, height)
+    
+    return this.attr('d', this.array.size(p.width, p.height))
   }
   // Set width of element
 , width: function(width) {
-    var b = this.bbox()
-
-    return width == null ? b.width : this.size(width, b.height)
+    return width == null ? this.bbox().width : this.size(width, this.bbox().height)
   }
   // Set height of element
 , height: function(height) {
-    var b = this.bbox()
+    return height == null ? this.bbox().height : this.size(this.bbox().width, height)
+  }
 
-    return height == null ? b.height : this.size(b.width, height)
-  }
-  // Set the actual size in pixels
-, size: function(width, height) {
-    var scale = width / this._offset.width
-    
-    return this.transform({
-      scaleX: scale
-    , scaleY: height != null ? height / this._offset.height : scale
-    })
-  }
-  // Set path data
-, plot: function(data) {
-    var x = this.trans.scaleX
-      , y = this.trans.scaleY
-    
-    /* native plot */
-    this._plot(data)
-    
-    /* store offset */
-    this._offset = this.transform({ scaleX: 1, scaleY: 1 }).bbox()
-    
-    /* get and store the actual offset of the element */
-    if (this.unbiased) {
-      this._offset.x = this._offset.y = 0
-    } else {
-      this._offset.x -= this.trans.x
-      this._offset.y -= this.trans.y
-    }
-    
-    return this.transform({ scaleX: x, scaleY: y })
-  }
-  // Private: Native plot
-, _plot: function(data) {
-    return this.attr('d', data || 'M0,0')
-  }
-  
 })
 
 //
 SVG.extend(SVG.Container, {
   // Create a wrapped path element
-  path: function(data, unbiased) {
-    return this.put(new SVG.Path(unbiased)).plot(data)
+  path: function(d) {
+    return this.put(new SVG.Path).plot(d)
   }
 
 })